# context/foundation/openapi.yaml
openapi: 3.1.0
...
info:
  title: 2do AI — API
  description: >
    Identity and session management for 2do AI (S-01): register an account, log in / out, read
    the current user, and permanently delete the account and all its data (FR-019). Also exposes
    the read-only list of the 11 fixed life domains (categories) the app navigation is built from,
    and the goals resource (S-02/S-07) — current tasks, long-term goals and someday dreams,
    per-user and full CRUD. The proposals resource (S-04) is the proactive half: it picks the one
    neglected entry worth coming back to and phrases it. Session model is a server-side HttpSession cookie (not JWT); mutations are
    CSRF-protected via a double-submit XSRF-TOKEN cookie echoed as the X-XSRF-TOKEN header.
  # #116 — MAJOR.MINOR.PATCH over the *contract*, not the app. 1.0.0 was S-01 (users, sessions,
  # categories); +goals (S-02) → 1.1.0, +proposals (S-04a) → 1.2.0, +current tasks (S-07) → 1.3.0,
  # +the phrased, stored, answerable proposal (S-04b) → 1.4.0, +the pending read the natural rhythm
  # made necessary (S-05) → 1.5.0, +the account language and the Accept-Language negotiation it is
  # the persistence for (DEV-49) → 1.6.0, +the category labels answering in that language (DEV-49)
  # → 1.7.0, +the two proposal operations declaring the `Accept-Language` they already honoured
  # (DEV-49) → 1.7.1 — a PATCH, because it documents behaviour the server already had rather than
...
  # (DEV-51) → 1.8.0.
  #
...
  version: 1.8.0
...
servers:
...
paths:
  /users:
    post:
      tags: [users]
      summary: Register a new user
      description: >
        Creates a user from an email + password. Does not log the user in (session creation is the
        sole responsibility of POST /sessions). The email is normalized (trimmed + lowercased) and
        must be unique. The new account's `language` is the one this
        request negotiated (FR-003) — the sign-up screen's language IS the Accept-Language the
        sign-up request carried, so there is deliberately no field for it on the payload.


        The account is created **unverified** and a six-digit code is emailed to the address, in the
        account's language and valid for 15 minutes: it cannot log in and the natural rhythm never
        writes to it until POST /verifications has been answered. An address whose account exists but
        was never verified is **taken over** — its password and language are replaced and a fresh code
        is sent, so an abandoned or mistyped sign-up never blocks the real owner — so only a *verified*
        duplicate is 409. The code is mailed after the account is committed, which is why a mail
        outage is 503 on an account that nevertheless exists; retry, or ask for a code again.
      operationId: createUser
      security: []                                          # public (#104 exception: registration)
      parameters:
        - $ref: '#/components/parameters/XsrfToken'         # public, but still a mutation — CSRF applies
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User created.
          headers:
            Location:
              description: Canonical location of the created user (readable once authenticated).
              schema:
                type: string
                format: uri-reference
                example: /api/users/me
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'          # missing / invalid CSRF token
        '409':
          $ref: '#/components/responses/Conflict'           # a *verified* account already has this email
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'    # too many codes to this address already
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          # 3.1 lets a Reference Object override the target's description.
          $ref: '#/components/responses/ServiceUnavailable'
          description: >
            The account was created but the code could not be mailed. Not a failed registration: the
            row exists and is unverified, so the remedy is POST /verification-codes (or signing up
            again with the same address), not a different email.
  /users/me:
    get:
      tags: [users]
      summary: Get the current user
      description: >
        Returns the authenticated user. The SPA calls this on load to bootstrap auth state; the
        response (or its 401) also carries the XSRF-TOKEN cookie primed by the server so the first
        mutation can echo it.
      operationId: getCurrentUser
      responses:
        '200':
          description: The current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [users]
      summary: Change the current user's language (FR-002)
      description: >
        Sets the language this account reads in. PATCH rather than PUT because the other members of
        `User` are the id and the email, neither of which this operation may touch. The change takes
        effect on the *calling session* immediately — the response is the updated user, and the very
        next GET /users/me reports the new value without a re-login.
      operationId: updateCurrentUser
      parameters:
        - $ref: '#/components/parameters/XsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: The updated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'         # a `language` outside the enum never deserializes
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'          # missing / invalid CSRF token
        '422':
          $ref: '#/components/responses/UnprocessableEntity' # `language` omitted
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [users]
      summary: Delete the current user and all their data (FR-019)
      description: >
        Permanently removes the authenticated user and every per-user record they own (their
        ai_memory profile facts + episodes), then invalidates the session. Irreversible. Requires
        re-entering the password as a re-authentication step. NOTE: a request body on DELETE has no
        defined HTTP semantics (RFC 9110); it works same-origin behind Cloudflare, and is used here
        so the re-auth password travels with the delete rather than in a separate step.
      operationId: deleteCurrentUser
      parameters:
        - $ref: '#/components/parameters/XsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDeletion'
      responses:
        '204':
          description: Account and all data deleted; session invalidated.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'       # not logged in / session expired — never a wrong password
        '403':
          # 3.1 lets a Reference Object override the target's description, so the shared Forbidden
          # response stays the single definition of the media type and schema.
          $ref: '#/components/responses/Forbidden'
          description: >
            Two distinct causes, told apart by the Problem `type` (never by parsing `detail`):
            `urn:2doai:problem:re-auth-failed` means the confirmation password is wrong — the session
            is intact and the user may retry. Any other `type`, **including its absence** (the server
            omits the member entirely on this path), is a missing or invalid CSRF token, which only a
            page reload re-primes, so a client must not offer a bare retry as the remedy for it.
            A wrong password is deliberately NOT a 401: on an authenticated call every SPA reads 401
            as "session expired" and would log the user out over a typo.
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /verifications:
    post:
      tags: [verifications]
      summary: Prove an email address with the code that was mailed to it
      description: >
        Spends the six-digit code POST /users (or POST /verification-codes) emailed to the address.
        On success the account becomes one that can log in and that the proactive engine may write
        to; the code is cleared in the same write, so it cannot be spent twice.


        Does NOT open a session — POST /sessions remains the sole session-creation path, so the
        client logs in afterwards.


        Every failure is one 403 with `type: urn:2doai:problem:verification-failed` and one fixed
        `detail`: unknown address, already verified, no outstanding code, expired, out of attempts,
        and simply wrong are deliberately indistinguishable. Telling them apart would say which
        addresses have unconfirmed accounts, and none of the six leaves the user anything different
        to do — ask for another code. A code is worth five wrong guesses; a fresh one restores the
        budget.
      operationId: createVerification
      security: []                                          # public (the account cannot log in yet)
      parameters:
        - $ref: '#/components/parameters/XsrfToken'         # public, but still a mutation — CSRF applies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerification'
      responses:
        '204':
          description: The address is confirmed; the account may now log in.
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
          description: >
            `urn:2doai:problem:verification-failed` — the code is wrong or no longer valid (see the
            operation description; the causes are deliberately conflated). Any other `type`,
            including its absence, is a missing or invalid CSRF token.
        '422':
          $ref: '#/components/responses/UnprocessableEntity' # `code` is not six digits
        '500':
          $ref: '#/components/responses/InternalServerError'
  /verification-codes:
    post:
      tags: [verifications]
      summary: Send another verification code to an address
      description: >
        Issues a fresh code for an unverified account, in that account's language, and invalidates
        the outstanding one — so a resend never widens the window a guesser is working in. The new
        code also restores the five attempts the old one may have spent.


        **202 for every address**, including one with no account and one already verified, neither of
        which is sent anything. "Accepted" is the honest word: the server has taken the request, and
        whether a message follows is exactly what it declines to disclose. The throttle is consulted
        before the address is looked up for the same reason — a 429 that only ever happened to real
        accounts would answer the question the 202 refuses to.
      operationId: createVerificationCode
      security: []                                          # public (the account cannot log in yet)
      parameters:
        - $ref: '#/components/parameters/XsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationCodeRequest'
      responses:
        '202':
          description: >
            The request was accepted. Says nothing about whether the address has an account, whether
            it was already verified, or whether a message was sent.
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'          # missing / invalid CSRF token
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable' # the provider would not take the message
...
components:
  securitySchemes:
    sessionCookie:
      # Server-side HttpSession, referenced by an opaque cookie (Spring's JSESSIONID). Not a bearer
      # token — validation is in-memory on the single Fly machine and never queries the database.
      type: apiKey
      in: cookie
      name: JSESSIONID

  # Every operation is authenticated by default; the public entry points override with
  # `security: []` (register, login, ping).
  # (applied at the operation level above via `security: []`; global default below)
  #
  # #105/#225 (permissions and their naming) are deliberately not implemented, and this is the note
  # that says so rather than leaving the omission to look like an oversight. Scopes exist to let one
  # API grant different subsets of itself to different callers; here there is exactly one caller (the
  # SPA, `x-audience: component-internal`) and every operation acts on the caller's own data. The
  # authorization rule is ownership, enforced per query in the service layer — a scope model on top
  # would be a second vocabulary describing the same single permission. Revisit when a second client
  # appears, or when any operation reads data the caller does not own (an admin or export path).
  parameters:
    XsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: >
        CSRF token echoed from the XSRF-TOKEN cookie on every mutating request (double-submit).
        (#132 would spell this X-Xsrf-Token; the functional header Spring reads is X-XSRF-TOKEN.)
      schema:
        type: string
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: >
        The language to answer in. This is the *transport* for the app's locale; `User.language` is
        the persistence. The SPA sends whatever it is currently rendering on every call, so the
        server never has to load an account to know what to write — which is what keeps
        GET /users/me free of any database query. Negotiation is the standard quality-ordered one:
        the first supported tag in the list wins, region is ignored (`en-GB` is `en`), and a header
        naming neither supported language — or no header at all — means English. Sent by the SPA on
        every request; only the operations that generate text for a person read it.
      schema:
        type: string
        example: en-GB,en;q=0.9
...
  schemas:
    User:
      type: object
      description: A user identity. Minimal by design (YAGNI) — display name / flags come later.
      required: [id, email, language]
      properties:
        id:
          type: string
          format: uuid                                      # UUID v7 surrogate PK (intentional, #144)
          readOnly: true
        email:
          type: string
          format: email
          example: alice@example.com
        language:
          $ref: '#/components/schemas/AppLanguage'
    AppLanguage:
      # THE ANCHOR for these two literals, the way GoalLayer is for its three — named once and
      # $ref'd by User and UserUpdate rather than spelled out in each.
      type: string
      description: >
        The language a person is addressed in. Two letters, not a BCP 47 tag: there are exactly two
        locales and no region variants, so a client choosing between `en` and `en-GB` would be
        choosing about nothing. This is what the account reads in, and what the SPA seeds itself
        from at login — before it, the browser decides what the auth screens are rendered in; after
        it, the account does. An account created before this field existed reports PL: null in
        storage means "never chosen", and those accounts were never asked (FR-004).
      # #112 open-ended enum: a third language must not break existing clients.
      x-extensible-enum:
        - PL
        - EN
      example: EN
    UserUpdate:
      type: object
      description: The account's language setting (FR-002). See PATCH /users/me.
      required: [language]
      properties:
        language:
          $ref: '#/components/schemas/AppLanguage'
    UserRegistration:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
          example: alice@example.com
        password:
          type: string
          format: password
          minLength: 8                                      # baseline hardening; mirrors @Size(min=8)
          maxLength: 72
          x-max-utf8-bytes: 72
          description: At most 72 UTF-8 bytes (the BCrypt input limit).
          writeOnly: true
...
    EmailVerification:
      type: object
      description: The code a user typed, for POST /verifications.
      required: [email, code]
      properties:
        email:
          type: string
          maxLength: 320
          description: >
            Validated as loosely as `Credentials.email`, and for the same reason: this operation
            answers identically for every address it does not recognise, and a stricter constraint
            would start answering 422 for some of them and 403 for others.
        code:
          type: string
          pattern: '^\\d{6}$'
          description: >
            Exactly six digits. A malformed code is 422 and never spends one of the five guesses the
            real code is worth.
          example: '042317'
    VerificationCodeRequest:
      type: object
      description: >
        The address to send another code to, for POST /verification-codes. Address only — which
        account it belongs to, and whether it has one at all, is deliberately not something the
        answer reveals.
      required: [email]
      properties:
        email:
          type: string
          maxLength: 320
          example: alice@example.com
    AccountDeletion:
      type: object
      description: Re-authentication for the irreversible FR-019 account deletion.
      required: [password]
      properties:
        password:
          type: string
          format: password
          maxLength: 72
          x-max-utf8-bytes: 72
          description: At most 72 UTF-8 bytes (the BCrypt input limit).
          writeOnly: true
...
    Problem:
      # RFC 9457 Problem Details (#176). Emitted as application/problem+json for every error.
      type: object
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
  responses:
    BadRequest:
      description: Malformed request syntax.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authenticated but the request is not allowed (e.g. missing CSRF token).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Conflict:
      description: Conflicting state (e.g. the email is already registered).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnprocessableEntity:
      description: The request was well-formed but failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    TooManyRequests:
      description: >
        Too many verification codes have been asked for this address — one per minute and five per
        hour. `Retry-After` carries the number of seconds to wait; it is the only part of this
        response a client can act on, so a screen must read it rather than guess. The limit is per
        address and held in the server's memory, so it resets on a restart.
      headers:
        Retry-After:
          description: Seconds to wait before asking again.
          required: true
          schema:
            type: integer
            format: int32
            example: 42
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalServerError:
      description: >
        An unhandled server fault. Reachable on every operation — `ApiExceptionHandler` and
        `ProblemErrorController` render it as Problem JSON like any other error, so a client parses
        it the same way rather than falling into an unrecognized-status branch.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ServiceUnavailable:
      description: >
        A dependency would not answer, and nothing the caller sent is wrong. On POST /sessions:
        authentication could not be completed for an infrastructure reason (a connection-pool
        timeout, a database cold start that overran) — deliberately NOT reported as 401, which would
        tell every user their password is wrong during an outage; the credentials were never judged.
        On the operations that email a code: the provider would not take the message. Retry in both
        cases.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

# Authenticated by default (#104); register + login opt out with `security: []`.
security:
  - sessionCookie: []